Factorize generation of Indian and Chinese SSNs
